app-backend/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
argcomplete==1.10.3
aspose-words==22.5.0
astunparse==1.6.3
autocorrect==2.6.1
beautifulsoup4==4.8.2
blis==0.7.5
cachetools==4.2.1
catalogue==2.0.6
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
ci-info==0.3.0
click==8.0.4
colorama==0.4.4
compressed-rtf==1.0.6
configobj==5.0.8
configparser==5.3.0
contourpy==1.1.0
cryptography==37.0.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.23
distlib==0.3.2
distro==1.6.0
dnspython==2.3.0
docx2pdf==0.1.8
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
etelemetry==0.3.0
extract-msg==0.28.7
filelock==3.0.12
flashtext==2.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.40.0
future==0.18.2
gast==0.4.0
gensim==4.1.2
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.56.0
googletrans==3.1.0a0
greenlet==1.1.2
grpcio==1.34.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.12.1
http-ece==1.1.0
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.0.17
hyperframe==5.2.0
idna==2.10
IMAPClient==2.1.0
importlib-metadata==6.7.0
importlib-resources==6.0.0
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
looseversion==1.2.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mars==0.0.0
marshmallow==3.11.1
matplotlib==3.6.0
murmurhash==1.0.6
networkx==2.6.3
nibabel==5.1.0
nipype==1.8.6
nltk==3.2.5
numpy==1.19.5
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
paramiko==3.0.0
pathlib==1.0.1
pathy==0.6.1
pdf-wrangler==0.0.25
pdfminer.six==20181108
Pillow==8.3.2
preshed==3.0.6
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.15.8
prov==2.0.0
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.8.2
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.4.0
pymongo==4.0.1
PyMuPDF==1.19.6
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF3==1.0.5
pypng==0.20220715.0
PySocks==1.7.1
pyspellchecker==0.6.3
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2021.3
pywebpush==1.14.0
pywin32==303
pywsd==1.0.2
pyxnat==1.5
PyYAML==5.4.1
qrcode==7.4.2
questionary==1.10.0
rdflib==6.3.2
regex==2021.9.30
requests==2.31.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==11.2.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
simplejson==3.19.1
six==1.15.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy-legacy==3.0.9
spacy-loggers==1.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.28
srsly==2.4.3
tabula==1.0.5
tabula-py==2.3.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.5.0
tensorflow-metadata==1.7.0
termcolor==1.1.0
textract==1.5.0
thinc==8.0.17
threadpoolctl==3.0.0
threads-net==0.0.49
tokenizers==0.10.3
torch==1.9.1
tqdm==4.62.3
traits==6.3.2
transformers==4.11.2
twitter==1.19.6
typeguard==2.13.3
typer==0.4.0
typing-extensions==3.7.4.3
tzdata==2021.2.post0
tzlocal==3.0
Unidecode==1.3.2
urllib3==1.26.4
virtualenv==20.4.7
waitress==2.1.0
wasabi==0.9.1
wcwidth==0.2.6
Werkzeug==2.0.2
wrapt==1.12.1
xlrd==1.2.0
zipp==3.15.0